+++ ScreenRecorderLibNative/HighresTimer.h
@@ -9,17 +9,17 @@ public:
 	HighresTimer();
 	~HighresTimer();
 	inline HANDLE GetTickEvent() { return m_TickEvent; }
-	HRESULT StartRecurringTimer(UINT msInterval);
+	HRESULT StartRecurringTimer(INT64 msInterval);
 	HRESULT StopTimer(bool waitForCompletion);
 	HRESULT WaitForNextTick();
-	HRESULT WaitFor(UINT64 interval100Nanos);
+	HRESULT WaitFor(INT64 interval100Nanos);
 	double GetMillisUntilNextTick();
-	inline UINT64 GetTickCount() { return m_TickCount; }
+	inline INT64 GetTickCount() { return m_TickCount; }
 private:
 	bool m_IsActive;
-	UINT64 m_TickCount;
+	INT64 m_TickCount;
 	std::chrono::steady_clock::time_point m_LastTick;
-	UINT m_Interval;
+	INT64 m_Interval;
 	UINT m_TimerResolution;
 	HANDLE m_TickEvent;
 	HANDLE m_StopEvent;
